[REF] statement_import_sheet_file: Allow to define more parsers
