+++ account_statement_import_sheet_file/readme/CONTRIBUTORS.md
@@ -12,3 +12,4 @@
   - Alexey Pelykh \<<alexey.pelykh@corphub.eu>\>
 - Sebastiano Picchi <sebastiano.picchi@pytech.it>
 - Jacques-Etienne Baudoux (BCIM) <je@bcim.be>
+- Simone Rubino
